# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Mystics Inventory API — multi-tenant inventory SaaS
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /team/users:
    post:
      operationId: createTeamUser
      tags: [team]
      summary: Create a new user account and add them to this organization.
      description: |
        Owners and admins can use this to create a teammate account
        directly with a password (skipping the invitation email
        round-trip). The new account is marked email-verified
        immediately, so the user can sign in straight away.

        If the email already belongs to an existing user (in any
        organization), the request is rejected with 409. The admin
        should use the invitation flow in that case so the existing
        user can opt in instead of having their password silently
        overwritten.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateTeamUserPayload" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TeamMember" }
        "403":
          description: Caller is not an owner/admin (or is not an owner trying to create another owner).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "409":
          description: A user with that email already exists.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /team/invitations:
    post:
      operationId: createTeamInvitation
      tags: [team]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateTeamInvitationPayload" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TeamInvitation" }
...
  /team/invitations/{id}:
    delete:
      operationId: revokeTeamInvitation
      tags: [team]
      parameters:
        - { in: path, name: id, required: true, schema: { type: integer } }
      responses:
        "204": { description: No Content }
  /team/invitations/accept:
    post:
      operationId: acceptTeamInvitation
      tags: [team]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AcceptInvitationPayload" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TeamMember" }
  /team/members/{id}:
    patch:
      operationId: updateTeamMemberRole
      tags: [team]
      parameters:
        - { in: path, name: id, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateTeamMemberRolePayload" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TeamMember" }
    delete:
      operationId: removeTeamMember
      tags: [team]
      parameters:
        - { in: path, name: id, required: true, schema: { type: integer } }
      responses:
        "204": { description: No Content }
  /team/members/{id}/permissions:
    patch:
      operationId: updateTeamMemberPermissions
      tags: [team]
      parameters:
        - { in: path, name: id, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateTeamMemberPermissionsPayload" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TeamMember" }
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        error: { type: string }
        code: { type: string }
      required: [error]
...
    TeamMember:
      type: object
      properties:
        id: { type: integer }
        userId: { type: integer }
        email: { type: string }
        name: { type: ["string", "null"] }
        role: { type: string }
        canEditBills: { type: boolean }
        canEditStocks: { type: boolean }
        createdAt: { type: string }
      required: [id, userId, email, name, role, canEditBills, canEditStocks, createdAt]
    TeamInvitation:
      type: object
      properties:
        id: { type: integer }
        email: { type: string }
        role: { type: string }
        token: { type: string }
        expiresAt: { type: string }
        acceptedAt: { type: ["string", "null"] }
        createdAt: { type: string }
      required: [id, email, role, token, expiresAt, acceptedAt, createdAt]
    CreateTeamInvitationPayload:
      type: object
      properties:
        email: { type: string }
        role: { type: string }
      required: [email, role]
    AcceptInvitationPayload:
      type: object
      properties:
        token: { type: string }
      required: [token]
    UpdateTeamMemberRolePayload:
      type: object
      properties:
        role: { type: string }
      required: [role]
    UpdateTeamMemberPermissionsPayload:
      type: object
      properties:
        canEditBills: { type: boolean }
        canEditStocks: { type: boolean }
...
    CreateTeamUserPayload:
      type: object
      properties:
        username: { type: string, minLength: 3, maxLength: 30 }
        email: { type: string }
        name: { type: string }
        password: { type: string, minLength: 8 }
        role: { type: string }
      required: [username, email, name, password, role]
...