# lib/api-zod/src/generated/types/createTeamUserPayload.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Mystics Inventory API — multi-tenant inventory SaaS
 * OpenAPI spec version: 0.1.0
 */

export interface CreateTeamUserPayload {
  /**
   * @minLength 3
   * @maxLength 30
   */
  username: string;
  email: string;
  name: string;
  /** @minLength 8 */
  password: string;
  role: string;
}