# lib/api-zod/src/generated/types/teamMember.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Mystics Inventory API — multi-tenant inventory SaaS
 * OpenAPI spec version: 0.1.0
 */

export interface TeamMember {
  id: number;
  userId: number;
  email: string;
  /** @nullable */
  name: string | null;
  role: string;
  canEditBills: boolean;
  canEditStocks: boolean;
  createdAt: string;
}